remove (wrong) magic number and improve comments
The photo byte size was set with a magic number that was actually wrong,
overshooting the end of the file.
dd(1) actually disregards this and doesn't read over the end of the file,
so the magic value didn't have any meaning or effect.

It is a safe bet to get rid of it.

Signed-off-by: Nuno Goncalves <nunojpg@gmail.com>